[traffic_logger] Add an initial traffic logger companion application
The traffic logger provides an API endpoint for services to log utilization
with the application. The actual network packet examination code will
be implemented in a faster language to support line-rate packet
processing, and will aggregate information to report regularly to the
traffic logger endpoint. Right now the logger lacks security and auth
protections, and should be confined to access only from the local host.