Allow domain names to be 255 characters in length
The DNS specification allows for domains of up to 253 characters, with
a stop and start byte for a total length of 255. To provide a "binary
round" size, set the max length to 255 even though we will probably
only ever see hosts of length 253 or less. Some observed labels are
already longer than the 50 characters currently allotted.