settings: Configure via instance settings via local config file
There is a difference between settings that are project wide and
settings that need to be customized for a specific
deployment. Separate the per-deployment settings into a TOML config
file to logically separate them from project-wide settings. This
commit creates a default config file for local development that is
used as a fallback if a system config is not available. It also
creates a hacky script to generate minimum system configs for
production.