[models] Update model names for clarity
With a better idea of how the commgestion webapp will interface with
haulage-rs, this commit renames and reorganizes some of the usage
related models for clarity. Due to the different time scales and
aggregation privacy constraints, each different type of utilization
will need to be tracked and logged separately. Logging them together
and then filtering will leave the underlying data vulnerable since it
still would exist in the community database.