[web/views] Add api to get the current utilization
This commit adds a view to get the current usage. The view is designed
as an API, returning json rather than a rendered HTML view. There may
be a need to move to a real api framework in the future (Django REST
framework?) but this is implemented directly in base django for now.