[web] Split chart render from data update
This commit splits the drawing of the gauge chart from the background
query for the gauge data and updating the chart. This split makes the
page more responsive (rendering the chart while the data query is
processing) and also lays the foundation to allow for polling of
updated query information.