[web] Remove public unit selection
The unit selection seems like it would likely be confusing to normal
users, or at least just inconvenient. Remove it for now to make more
space in the page. This should be replaced by an admin configuration
option to set the overall network throughput and the units used in the
UI consistently across pages.